Create test 0001 and constants
